hackathon_genius/sub_agents/idea_agent.py
"""
IdeaAgent — krok 1 pipeline'u Hackathon Genius.

Odbiera od użytkownika temat hackathonu i generuje 3 kreatywne pomysły
na projekt. Korzysta z narzędzia browse_hackathon_themes, aby pobrać
inspiracje z wcześniejszych zwycięskich projektów. Zapisuje wynik do
stanu sesji pod kluczem 'hackathon_ideas', aby kolejny agent mógł go
odczytać.
"""

from google.adk.agents import LlmAgent

from ..model_config import build_demo_model
from ..tools.hackathon_tools import browse_hackathon_themes

idea_agent = LlmAgent(
    name="IdeaAgent",
    model=build_demo_model(),
    description=(
        "Generuje 3 kreatywne i innowacyjne pomysły na projekt hackathonowy "
        "na podstawie podanego tematu."
    ),
    instruction="""Jesteś kreatywnym agentem do generowania pomysłów : 
    hackathonowych 
    lub projektow dla studentow informatyki na zaliczenie programowania.
Masz talent do dostrzegania ciekawych szans produktowych i technologicznych.

Zawsze odpowiadaj wyłącznie po polsku.
Nawet jeśli narzędzie zwróci dane po angielsku, końcową odpowiedź sformatuj po polsku.

Twoje zadanie:
1. Weź temat hackathonu podany przez użytkownika.
2. Wywołaj narzędzie `browse_hackathon_themes` z tym tematem, aby pobrać inspiracje z poprzednich zwycięskich projektów.
3. Na podstawie wyniku narzędzia ORAZ własnej kreatywności wygeneruj dokładnie 3 unikalne pomysły na projekt.

Dla KAŻDEGO pomysłu podaj:
- 💡 **Nazwa projektu**: chwytliwy i zapamiętywalny tytuł
- 🎯 **Rozwiązywany problem**: jedno zdanie opisujące realny problem użytkownika
- 🚀 **Kluczowa funkcja**: najważniejsza rzecz, którą robi produkt
- 🌟 **Efekt wow**: co wyróżnia ten pomysł na tle istniejących rozwiązań

Każdy pomysł ma być zwięzły, ale ekscytujący. Myśl o tym, co zrobi wrażenie na jury hackathonu lub nauczyciela przedmiotu ktory zadal to zadanie.
Użyj numerowanej listy: Pomysł 1, Pomysł 2, Pomysł 3.

Po przedstawieniu 3 pomysłów dodaj krótką linię:
"Przekazuję pomysły do agenta TechStackAgent..."
""",
    tools=[browse_hackathon_themes],
    output_key="hackathon_ideas",
)
