requirements.txt
python-dotenv
google-adk==1.19.0
litellm
pytest==9.0.1
pytest-asyncio==1.3.0
pandas